GameModel 파일 분리
